flytt behandlingUtils inn i utils mappe
Flyttet eq funksjoner nærmere der dem blir brukt
