+++ src/Utils/behandling/behandlingUtils.ts
@@ -1,5 +1,4 @@
 import * as Arr from 'fp-ts/Array';
-import { Eq } from 'fp-ts/lib/Eq';
 import { pipe } from 'fp-ts/lib/function';
 import * as Option from 'fp-ts/Option';
 
@@ -8,23 +7,14 @@ import {
     Vilkårsinformasjon,
     vilkårsinformasjonForBeregningssteg,
 } from '~features/saksoversikt/utils';
-import { Nullable } from '~lib/types';
 import { Behandling, Behandlingsstatus } from '~types/Behandling';
 import {
-    FastOppholdINorge,
     FastOppholdINorgeStatus,
-    Flyktning,
     FlyktningStatus,
-    Formue,
     FormueStatus,
-    FormueVerdier,
-    Institusjonsopphold,
     InstitusjonsoppholdStatus,
-    LovligOpphold,
     LovligOppholdStatus,
-    OppholdIUtlandet,
     OppholdIUtlandetStatus,
-    PersonligOppmøte,
     PersonligOppmøteStatus,
     UførhetStatus,
 } from '~types/Behandlingsinformasjon';
@@ -45,13 +35,6 @@ export function erIverksatt(behandling: Behandling): boolean {
     return [Behandlingsstatus.IVERKSATT_AVSLAG, Behandlingsstatus.IVERKSATT_INNVILGET].some(
         (status) => behandling.status === status
     );
-}
-export function erIverksattAvslag(behandling: Behandling): boolean {
-    return behandling.status === Behandlingsstatus.IVERKSATT_AVSLAG;
-}
-
-export function erIverksattInnvilget(behandling: Behandling): boolean {
-    return behandling.status === Behandlingsstatus.IVERKSATT_INNVILGET;
 }
 
 export function erAvslått(behandling: Behandling): boolean {
@@ -126,75 +109,4 @@ export const hentSisteVurdertSaksbehandlingssteg = (behandling: Behandling) => {
             (x) => x.vilkårtype
         )
     );
-};
-
-export const eqFlyktning: Eq<Nullable<Flyktning>> = {
-    equals: (flyktning1, flyktning2) =>
-        flyktning1?.status === flyktning2?.status && flyktning1?.begrunnelse === flyktning2?.begrunnelse,
-};
-
-export const eqLovligOppholdINorge: Eq<Nullable<LovligOpphold>> = {
-    equals: (lovlig1, lovlig2) => lovlig1?.status === lovlig2?.status && lovlig1?.begrunnelse === lovlig2?.begrunnelse,
-};
-
-export const eqFastOppholdINorge: Eq<Nullable<FastOppholdINorge>> = {
-    equals: (fastOpphold1, fastOpphold2) =>
-        fastOpphold1?.status === fastOpphold2?.status && fastOpphold1?.begrunnelse === fastOpphold2?.begrunnelse,
-};
-
-export const eqInstitusjonsopphold: Eq<Nullable<Institusjonsopphold>> = {
-    equals: (institusjonsopphold1, institusjonsopphold2) =>
-        institusjonsopphold1?.status === institusjonsopphold2?.status &&
-        institusjonsopphold1?.begrunnelse === institusjonsopphold2?.begrunnelse,
-};
-
-export const eqOppholdIUtlandet: Eq<Nullable<OppholdIUtlandet>> = {
-    equals: (oppholdIUtlandet1, oppholdIUtlandet2) =>
-        oppholdIUtlandet1?.status === oppholdIUtlandet2?.status &&
-        oppholdIUtlandet1?.begrunnelse === oppholdIUtlandet2?.begrunnelse,
-};
-
-export const eqFormue: Eq<Nullable<Formue>> = {
-    equals: (formue1, formue2) =>
-        formue1?.status === formue2?.status &&
-        eqVerdier.equals(formue1?.verdier ?? null, formue2?.verdier ?? null) &&
-        eqVerdier.equals(formue1?.epsVerdier ?? null, formue2?.epsVerdier ?? null) &&
-        formue1?.begrunnelse === formue2?.begrunnelse,
-};
-const eqVerdier: Eq<Nullable<FormueVerdier>> = {
-    equals: (verdier1, verdier2) =>
-        verdier1?.verdiIkkePrimærbolig === verdier2?.verdiIkkePrimærbolig &&
-        verdier1?.verdiKjøretøy === verdier2?.verdiKjøretøy &&
-        verdier1?.innskudd === verdier2?.innskudd &&
-        verdier1?.verdipapir === verdier2?.verdipapir &&
-        verdier1?.pengerSkyldt === verdier2?.pengerSkyldt &&
-        verdier1?.kontanter === verdier2?.kontanter &&
-        verdier1?.depositumskonto === verdier2?.depositumskonto,
-};
-
-export const eqEktefelle: Eq<
-    Nullable<{
-        fnr: Nullable<string>;
-    }>
-> = {
-    equals: (ektefelle1, ektefelle2) => ektefelle1?.fnr === ektefelle2?.fnr,
-};
-
-export const eqPersonligOppmøte: Eq<Nullable<PersonligOppmøte>> = {
-    equals: (personligOppmøte1, personligOppmøte2) =>
-        personligOppmøte1?.status === personligOppmøte2?.status &&
-        personligOppmøte1?.begrunnelse === personligOppmøte2?.begrunnelse,
-};
-
-export const eqBosituasjon: Eq<
-    Nullable<{
-        delerBolig: Nullable<boolean>;
-        ektemakeEllerSamboerUførFlyktning: Nullable<boolean>;
-        begrunnelse: Nullable<string>;
-    }>
-> = {
-    equals: (sats1, sats2) =>
-        sats1?.delerBolig === sats2?.delerBolig &&
-        sats1?.ektemakeEllerSamboerUførFlyktning === sats2?.ektemakeEllerSamboerUførFlyktning &&
-        sats1?.begrunnelse === sats2?.begrunnelse,
 };
