+++ src/components/søknadsbehandlingoppsummering/SøknadsbehandlingHeader.tsx
@@ -10,11 +10,11 @@ import { IntlShape } from 'react-intl';
 
 import * as PdfApi from '~api/pdfApi';
 import { useUserContext } from '~context/userContext';
-import { erIverksatt } from '~features/behandling/behandlingUtils';
 import { useI18n, useBrevForhåndsvisning } from '~lib/hooks';
 import { søknadMottatt } from '~lib/søknadUtils';
 import { Behandling, Behandlingsstatus, UnderkjennelseGrunn } from '~types/Behandling';
 import { Vedtak } from '~types/Vedtak';
+import { erIverksatt } from '~Utils/behandling/behandlingUtils';
 
 import messages from './søknadsbehandling-nb';
 import styles from './søknadsbehandlingHeader.module.less';
