+++ src/features/søknad/types.ts
@@ -1,10 +1,5 @@
 import { Nullable } from '~lib/types';
 
-export enum Bosituasjon {
-    BorAleneEllerMedBarnUnder18 = 'bor-alene-med-barn-under-18',
-    BorMedNoenOver18 = 'bor-med-noen-over-18',
-}
-
 export enum TypeOppholdstillatelse {
     Permanent = 'permanent',
     Midlertidig = 'midlertidig',
