+++ src/pages/saksbehandling/søknadsbehandling/fast-opphold-i-norge/FastOppholdINorge.tsx
@@ -1,5 +1,6 @@
 import * as RemoteData from '@devexperts/remote-data-ts';
 import { useFormik } from 'formik';
+import { Eq } from 'fp-ts/lib/Eq';
 import AlertStripe from 'nav-frontend-alertstriper';
 import { Feiloppsummering, Radio, RadioGruppe, Textarea } from 'nav-frontend-skjema';
 import NavFrontendSpinner from 'nav-frontend-spinner';
@@ -7,7 +8,6 @@ import React, { useState } from 'react';
 import { useHistory } from 'react-router-dom';
 
 import ToKolonner from '~components/toKolonner/ToKolonner';
-import { eqFastOppholdINorge } from '~features/behandling/behandlingUtils';
 import { lagreBehandlingsinformasjon } from '~features/saksoversikt/sak.slice';
 import { pipe } from '~lib/fp';
 import { useI18n } from '~lib/hooks';
@@ -29,6 +29,11 @@ interface FormData {
     status: Nullable<FastOppholdINorgeStatus>;
     begrunnelse: Nullable<string>;
 }
+
+const eqFastOppholdINorge: Eq<Nullable<FastOppholdINorgeType>> = {
+    equals: (fastOpphold1, fastOpphold2) =>
+        fastOpphold1?.status === fastOpphold2?.status && fastOpphold1?.begrunnelse === fastOpphold2?.begrunnelse,
+};
 
 const schema = yup.object<FormData>({
     status: yup
